requirements.txt
discord.py==2.3.2
PyNaCl==1.5.0
aiohttp==3.8.6
yt-dlp==2024.4.9
python-dotenv==1.0.0
psutil==5.9.8
mysql-connector-python==8.2.0
psycopg2-binary==2.9.9
youtube_dl==2021.12.17
PyYAML==6.0.1
requests==2.31.0
Pillow==10.1.0
numpy==1.26.2
matplotlib==3.8.2
beautifulsoup4==4.12.2
openai==1.3.8
googletrans==4.0.0rc1
vaderSentiment==3.3.2
wordcloud==1.9.2
plotly==5.17.0
pandas==2.1.4
seaborn==0.13.0
scikit-learn==1.3.2
flask==3.0.0
uvicorn==0.24.0
fastapi==0.104.1
redis==5.0.1
celery==5.3.4
sqlalchemy==2.0.23
alembic==1.12.1
pytest==7.4.3
black==23.11.0
flake8==6.1.0